fix: custom cron option issue and cron label issue
