+++ openmetadata-ui/src/main/resources/ui/src/utils/CronUtils.ts
@@ -12,14 +12,25 @@
  */
 
 import {
+  combinations,
   getDayCron,
   getHourCron,
   getMinuteCron,
   getMonthCron,
   getWeekCron,
   getYearCron,
+  SELECTED_PERIOD_OPTIONS,
+  toDisplay,
 } from 'components/common/CronEditor/CronEditor.constant';
-import { StateValue } from 'components/common/CronEditor/CronEditor.interface';
+import {
+  Combination,
+  CronType,
+  CronValue,
+  SelectedYearOption,
+  StateValue,
+  ToDisplay,
+} from 'components/common/CronEditor/CronEditor.interface';
+import { isEmpty, toNumber } from 'lodash';
 
 export const getCron = (state: StateValue) => {
   const {
@@ -48,4 +59,83 @@ export const getCron = (state: StateValue) => {
     default:
       return null;
   }
+};
+
+const getCronType = (cronStr: string) => {
+  for (const c in combinations) {
+    if (combinations[c as keyof Combination].test(cronStr)) {
+      return c;
+    }
+  }
+
+  return undefined;
+};
+
+export const getStateValue = (valueStr: string) => {
+  const stateVal: StateValue = {
+    selectedPeriod: '',
+    selectedMinOption: {
+      min: 5,
+    },
+    selectedHourOption: {
+      min: 0,
+    },
+    selectedDayOption: {
+      hour: 0,
+      min: 0,
+    },
+    selectedWeekOption: {
+      dow: 1,
+      hour: 0,
+      min: 0,
+    },
+    selectedMonthOption: {
+      dom: 1,
+      hour: 0,
+      min: 0,
+    },
+    selectedYearOption: {
+      dom: 1,
+      mon: 1,
+      hour: 0,
+      min: 0,
+    },
+  };
+  const cronType = getCronType(valueStr);
+
+  const d = valueStr ? valueStr.split(' ') : [];
+  const v: CronValue = {
+    min: d[0],
+    hour: d[1],
+    dom: d[2],
+    mon: d[3],
+    dow: d[4],
+  };
+
+  stateVal.selectedPeriod = cronType || stateVal.selectedPeriod;
+
+  if (!isEmpty(cronType)) {
+    const stateIndex =
+      SELECTED_PERIOD_OPTIONS[(cronType as CronType) || 'hour'];
+    const selectedPeriodObj = stateVal[
+      stateIndex as keyof StateValue
+    ] as SelectedYearOption;
+
+    const targets = toDisplay[cronType as keyof ToDisplay];
+
+    for (const element of targets) {
+      const tgt = element;
+
+      if (tgt === 'time') {
+        selectedPeriodObj.hour = toNumber(v.hour);
+        selectedPeriodObj.min = toNumber(v.min);
+      } else {
+        selectedPeriodObj[tgt as keyof SelectedYearOption] = toNumber(
+          v[tgt as keyof CronValue]
+        );
+      }
+    }
+  }
+
+  return stateVal;
 };
